+++ src/main/java/edu/uph/ii/platformy/repositories/StypendiaRepository.java
@@ -4,7 +4,7 @@ import edu.uph.ii.platformy.models.Specjalnosci;
 import edu.uph.ii.platformy.models.Stypendia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StypendiaRepository extends JpaRepository<Stypendia, Integer> {
+public interface StypendiaRepository extends JpaRepository<Stypendia, Long> {
 
     Stypendia findByName(String name);
 }
